+++ src/page/index.js
@@ -23,8 +23,7 @@ import FarmerProduct from "./FarmerProduct";
 import FarmerCreateProduct from "./FarmerCreateProduct";
 import CustomerHome from "./CustomerHome";
 import ProductDetail from "./ProductDetail";
-
-import CustomerCartView from "./CustomerCartView";
+import CustomerMyCart from "./CustomerMyCart";
 
 export {
   ProductDetail,
@@ -52,5 +51,5 @@ export {
   CustomerCreateTransaction,
   CustomerCreateTransactionLine,
   FarmerProduct,
-  CustomerCartView,
+  CustomerMyCart,
 };
